// src/payload-types.ts
export interface Category {
  id: string;
  name?: string | null;
  slug: string;
  color?: string | null;
  parent?: (string | null) | Category;
  subCategories?: {
    docs?: (string | Category)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  updatedAt: string;
  createdAt: string;
}